requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
gunicorn==23.0.0

# LLM Inference (CPU)
llama-cpp-python==0.3.4

# Embedding + Reranker (CrossEncoder)
sentence-transformers==3.3.1
FlagEmbedding==1.3.3

# Vision Model (optional, for PDF understanding)
transformers>=4.40.0
Pillow>=10.0.0

# Vector Store
qdrant-client==1.12.1

# PDF Processing
pymupdf==1.25.1
python-docx==1.1.2

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
trafilatura==2.0.0
html2text==2024.2.26

# Utilities
markdown==3.7
tiktoken==0.8.0
numpy==1.26.4
